!!! to be tested !!! -> buy / sell from player's shop, -> test pvp maps, -> test dod map (party pvp map?),
*** hack ***
-> adding some checks in vending functions,

*** Party ***
-> givin an ID to the party,
-> the ID is static and in sequence, like client ID, should be faster,

*** PVP ***
-> pvp maps with party should work ok now (to be tested),
-> using long int for pvp now rather than int,

+++ World Server/player.h
@@ -59,7 +59,7 @@ class CPlayer: public CCharacter
     bool isInvisibleMode;
 
     //LMA: Pvp flag
-    int pvp_id;
+    long int pvp_id;
 
     //LMA: Warp flag
     int map_warp_zone;
@@ -249,7 +249,7 @@ class CPlayer: public CCharacter
     bool VisiblityList( );
     bool ForceRefreshMonster(bool refresh_all, UINT monster_id);      //LMA: force refresh (20070621-211100)
     bool SpawnToPlayer( CPlayer* player, CPlayer* otherclient );
-    int ReturnPvp( CPlayer* player, CPlayer* otherclient );    //LMA: for PVP.
+    long int ReturnPvp( CPlayer* player, CPlayer* otherclient );    //LMA: for PVP.
     bool UpdateValues( );
     bool RefreshHPMP();  //LMA HP / MP Jumping
     void ReduceABC(int amount=1,bool do_packet=true);
